updated header files to work properly with new changes

+++ include/gameObjects.h
@@ -1,11 +1,11 @@
 #pragma once
-
+#include <cstddef>
 struct Enemy{
     int x;
     int y;
     int health;
     int strength;
-    int pathIndex;
+    size_t pathIndex;
     char symbol;
     bool alive; //1 alive, 0 dead
 };
